+++ qc-application/src/main/java/com/eager/questioncloud/application/api/cart/CartItemAppender.java
@@ -1,7 +1,7 @@
 package com.eager.questioncloud.application.api.cart;
 
-import com.eager.questioncloud.core.domain.cart.CartItem;
-import com.eager.questioncloud.core.domain.cart.CartItemRepository;
+import com.eager.questioncloud.core.domain.cart.infrastructure.CartItemRepository;
+import com.eager.questioncloud.core.domain.cart.model.CartItem;
 import com.eager.questioncloud.core.domain.question.QuestionRepository;
 import com.eager.questioncloud.core.domain.userquestion.UserQuestionRepository;
 import com.eager.questioncloud.exception.CustomException;
