+++ qc-application/src/main/java/com/eager/questioncloud/application/api/cart/CartResponse.java
@@ -1,6 +1,6 @@
 package com.eager.questioncloud.application.api.cart;
 
-import com.eager.questioncloud.core.domain.cart.CartItemDetail;
+import com.eager.questioncloud.core.domain.cart.dto.CartItemDetail;
 import java.util.List;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
