+++ qc-core/src/main/java/com/eager/questioncloud/core/domain/cart/model/CartItem.java
@@ -1,4 +1,4 @@
-package com.eager.questioncloud.core.domain.cart;
+package com.eager.questioncloud.core.domain.cart.model;
 
 import lombok.Builder;
 import lombok.Getter;
